Move Stripe price IDs to lib
